LayoutTests/media/modern-media-controls/ios-inline-media-controls/touch/ios-inline-media-controls-timeline-scrub-stops-propagation-expected.txt

Test that scrubbing the timeline does not propagate touch events to parent elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS containerReceivedTouchStart is false
PASS containerReceivedTouchMove is false
PASS containerReceivedTouchEnd is false
PASS successfullyParsed is true

TEST COMPLETE
